Move TopologyValidator into validators subpackage
